refactor: split each rule into its own module + UI fixture
Per review feedback (KSXGitHub/perfectionist#5):

- Move the `unicode_ellipsis_in_comments` lint into
  `src/unicode_ellipsis_in_comments.rs`. `src/lib.rs` becomes a
  scaffolding file: `extern crate` declarations, `dylint_library!()`,
  `mod <rule>;`, and a `register_lints` aggregator that calls each
  rule's `register(&mut LintStore)` function. Adding a new rule is
  now: drop a file in `src/`, add `mod <rule>;`, add one line to
  `register_lints`. No `mod.rs` per project convention.

- Rename `ui/main.rs` / `ui/main.stderr` to
  `ui/unicode_ellipsis_in_comments.{rs,stderr}`. `dylint_testing::ui_test`
  already discovers every `.rs` file in `ui/`, so each future rule
  drops its own fixture pair beside this one — no shared file to
  outgrow.

- Switch the config key from the package name (`perfectionist`) to
  the rule name (`unicode_ellipsis_in_comments`) so a second
  rule's config doesn't collide.
